Move post creation page component to `components/forms`
