+++ frontend/src/components/forms/create-post/CreatePostForm.tsx
@@ -1,28 +1,28 @@
 "use client"
 
-import { useRouter } from "next/navigation"
-import { Dispatch, FC, SetStateAction, useId, useRef, useState } from "react"
+import { FC, useRef, useState } from "react"
 import ReactTextAreaAutosize from "react-textarea-autosize"
-import Select from "react-select"
+import { useRouter } from "next/navigation"
 
-import { CreatePost, generateDescription } from "@/api/posts/create"
+import { CreatePost } from "@/api/posts/create"
 import { FormError } from "@/components/error"
-import { Capitalize } from "@/helpers/text"
 import { MarkdownToPlain } from "@/components/markdown/render"
 import MarkdownEditor from "@/components/markdown/editor"
 import { trimInput } from "@/helpers/input"
+import { GenerateDescriptionButton } from "@/components/forms/create-post/GenerateDescriptionButton"
+import SelectCategories from "@/components/forms/create-post/SelectCategories"
 
-const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
+const CreatePostForm: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
   categories,
   isAIEnabled,
 }) => {
-  const [formError, setFormError] = useState<string | null>(null)
-
-  const router = useRouter()
+  const formRef = useRef<HTMLFormElement>(null)
 
   const [isSame, setIsSame] = useState(false)
 
-  const formRef = useRef<HTMLFormElement>(null)
+  const [formError, setFormError] = useState<string | null>(null)
+
+  const router = useRouter()
 
   return (
     <form
@@ -43,8 +43,6 @@ const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
           description: formData.get("description") as string,
           categories: formData.getAll("categories") as string[],
         }
-
-        console.log(formFields)
 
         if (formError != null) setFormError(null)
         setIsSame(true)
@@ -72,6 +70,7 @@ const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
             <div className={"my-2 font-Yesteryear text-3xl text-primary opacity-50 text-center"}>
               <span className={"text-neutral-content"}> {"What's on your mind?"}</span>
             </div>
+
             <div className={"form-control"}>
               <input
                 onBlur={trimInput}
@@ -83,6 +82,7 @@ const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
                 maxLength={25}
               />
             </div>
+
             <div className={"form-control"}>
               <MarkdownEditor
                 title={
@@ -110,20 +110,11 @@ const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
             {isAIEnabled && (
               <GenerateDescriptionButton formRef={formRef} setFormError={setFormError} />
             )}
-            <div className={"mb-3"}>
-              <Select
-                placeholder={"Categories"}
-                instanceId={useId()}
-                isMulti={true}
-                name={"categories"}
-                className={"react-select-container"}
-                classNamePrefix={"react-select"}
-                options={categories.map((name) => ({ label: Capitalize(name), value: name }))}
-                required
-              />
-            </div>
+
+            <SelectCategories categories={categories} />
 
             <FormError error={formError} />
+
             <div className={"form-control"}>
               <button type={"submit"} className={"button self-center"}>
                 <svg
@@ -152,68 +143,4 @@ const CreatePostPage: FC<{ categories: string[]; isAIEnabled: boolean }> = ({
   )
 }
 
-const GenerateDescriptionButton: FC<{
-  formRef: React.RefObject<HTMLFormElement>
-  setFormError: Dispatch<SetStateAction<string | null>>
-}> = ({ formRef, setFormError }) => {
-  const [isLoading, setIsLoading] = useState(false)
-
-  return (
-    <button
-      onClick={async () => {
-        const formData = new FormData(formRef.current as HTMLFormElement)
-
-        const formFields = {
-          title: formData.get("title") as string,
-          content: formData.get("content") as string,
-        }
-
-        if (!formFields.title) {
-          return setFormError("Title is required")
-        }
-        if (!formFields.content) {
-          return setFormError("Content is required")
-        }
-        setIsLoading(true)
-        generateDescription(formFields)
-          .then((description) => {
-            setFormError(null)
-            const descriptionTextarea = formRef.current?.querySelector(
-              "[name=description]"
-            ) as HTMLInputElement
-            descriptionTextarea.value = description
-          })
-          .catch((err) => setFormError(err.message))
-          .finally(() => setIsLoading(false))
-      }}
-      type={"button"}
-      className={
-        "btn btn-sm transition-none hover:opacity-100 hover:gradient-text hover:border-primary btn-outline font-normal mb-3 self-center font-xs"
-      }
-    >
-      {isLoading ? (
-        <span className={"text-primary loading loading-ring"} />
-      ) : (
-        <svg
-          xmlns={"http://www.w3.org/2000/svg"}
-          fill={"none"}
-          viewBox={"0 0 24 24"}
-          strokeWidth={1}
-          stroke={"currentColor"}
-          className={"w-5 h-5 mr-1 fill-primary"}
-        >
-          <path
-            strokeLinecap={"round"}
-            strokeLinejoin={"round"}
-            d={
-              "M9.813 15.904L9 18.75l-.813-2.846a4.5 4.5 0 00-3.09-3.09L2.25 12l2.846-.813a4.5 4.5 0 003.09-3.09L9 5.25l.813 2.846a4.5 4.5 0 003.09 3.09L15.75 12l-2.846.813a4.5 4.5 0 00-3.09 3.09zM18.259 8.715L18 9.75l-.259-1.035a3.375 3.375 0 00-2.455-2.456L14.25 6l1.036-.259a3.375 3.375 0 002.455-2.456L18 2.25l.259 1.035a3.375 3.375 0 002.456 2.456L21.75 6l-1.035.259a3.375 3.375 0 00-2.456 2.456zM16.894 20.567L16.5 21.75l-.394-1.183a2.25 2.25 0 00-1.423-1.423L13.5 18.75l1.183-.394a2.25 2.25 0 001.423-1.423l.394-1.183.394 1.183a2.25 2.25 0 001.423 1.423l1.183.394-1.183.394a2.25 2.25 0 00-1.423 1.423z"
-            }
-          />
-        </svg>
-      )}
-      Generate description
-    </button>
-  )
-}
-
-export default CreatePostPage
+export default CreatePostForm
